keep the addNotificationReceivedListene in notification hook
