+++ blotztask-mobile/src/feature/badge/components/badge-achievement-modal.tsx
@@ -1,4 +1,4 @@
-import { useBadgeQueue } from "@/feature/badge/hooks/useBadgeQueue";
+import { BadgeNotificationDTO } from "@/feature/badge/models/badge-notification-dto";
 import { formatBadgeDate } from "@/feature/badge/utils/format-badge-date";
 import { GradientColor } from "@/shared/components/gradient-color";
 import MaterialIcons from "@react-native-vector-icons/material-icons/static";
@@ -6,24 +6,22 @@ import { Image } from "expo-image";
 import { useTranslation } from "react-i18next";
 import { Modal, Pressable, Text, View } from "react-native";
 
-export function BadgeAchievementModal() {
-  const { t } = useTranslation("badge");
-  const { currentBadge, dismissBadge } = useBadgeQueue();
+interface BadgeAchievementModalProps {
+  badge?: BadgeNotificationDTO;
+  onDismiss: () => void;
+}
 
-  if (!currentBadge) return null;
+export function BadgeAchievementModal({ badge, onDismiss }: BadgeAchievementModalProps) {
+  const { t } = useTranslation("badge");
+
+  if (!badge) return null;
 
   return (
-    <Modal
-      transparent
-      visible
-      animationType="fade"
-      statusBarTranslucent
-      onRequestClose={dismissBadge}
-    >
+    <Modal transparent visible animationType="fade" statusBarTranslucent onRequestClose={onDismiss}>
       <View className="flex-1 bg-black/50 items-center justify-center px-8">
         <View className="w-full items-center">
           <View className="w-full items-end mb-1">
-            <Pressable onPress={dismissBadge} hitSlop={12}>
+            <Pressable onPress={onDismiss} hitSlop={12}>
               <MaterialIcons name="highlight-off" size={30} color="#FFFFFF" />
             </Pressable>
           </View>
@@ -37,23 +35,23 @@ export function BadgeAchievementModal() {
           </View>
 
           <Image
-            source={{ uri: currentBadge.iconUrl }}
+            source={{ uri: badge.iconUrl }}
             style={{ width: 128, height: 128, marginBottom: 20 }}
             contentFit="contain"
           />
 
           <View className="h-14 bg-lime-100 rounded-full px-5 mb-2 items-center justify-center">
             <Text className="text-secondary text-2xl font-bold font-baloo text-center leading-8 pt-1">
-              {currentBadge.name}
+              {badge.name}
             </Text>
           </View>
 
           <Text className="text-white text-xl font-balooThin text-center w-64">
-            {currentBadge.description}
+            {badge.description}
           </Text>
 
           <Text className="text-white text-xl font-baloo mt-4">
-            {t("obtainedOn", { date: formatBadgeDate(currentBadge.obtainedAt) })}
+            {t("obtainedOn", { date: formatBadgeDate(badge.obtainedAt) })}
           </Text>
         </View>
       </View>
