+++ blotztask-mobile/src/shared/hooks/usePushNotificationSetup.ts
@@ -1,4 +1,5 @@
-import { useEffect } from "react";
+import { BadgeNotificationDTO } from "@/feature/badge/models/badge-notification-dto";
+import { useEffect, useState } from "react";
 import * as Notifications from "expo-notifications";
 import { registerForPushNotificationsAsync } from "@/shared/services/notifications";
 import { useRouter } from "expo-router";
@@ -6,9 +7,28 @@ import { toggleTaskCompletion } from "../services/task-service";
 
 export function usePushNotificationSetup() {
   const router = useRouter();
+  const [badgeQueue, setBadgeQueue] = useState<BadgeNotificationDTO[]>([]);
 
   useEffect(() => {
     registerForPushNotificationsAsync();
+
+    const notificationListener = Notifications.addNotificationReceivedListener((notification) => {
+      const data = notification.request.content.data;
+      if (data?.type !== "badge") return;
+
+      const earnedAtUtc = typeof data.earnedAtUtc === "string" ? data.earnedAtUtc : undefined;
+
+      setBadgeQueue((prev) => [
+        ...prev,
+        {
+          badgeId: Number(data.badgeId),
+          name: notification.request.content.body ?? "",
+          description: typeof data.description === "string" ? data.description : "",
+          iconUrl: typeof data.iconUrl === "string" ? data.iconUrl : "",
+          obtainedAt: earnedAtUtc ? new Date(earnedAtUtc) : new Date(),
+        },
+      ]);
+    });
 
     const responseListener = Notifications.addNotificationResponseReceivedListener(
       async (response) => {
@@ -29,7 +49,12 @@ export function usePushNotificationSetup() {
     );
 
     return () => {
+      notificationListener.remove();
       responseListener.remove();
     };
   }, []);
+
+  const dismissBadge = () => setBadgeQueue((prev) => prev.slice(1));
+
+  return { badgeQueue, dismissBadge };
 }
